CheckedSizeProduct

CheckedSizeProduct is a Julia package.

CheckedSizeProduct.StatusInvalidValue — Type

StatusInvalidValue::DataType

Singleton type. Returned by checked_size_product in case its input contained an invalid value. An invalid value is either:

less than zero

equal to typemax(T)

See also: StatusOverflow

CheckedSizeProduct.StatusOverflow — Type

StatusOverflow::DataType

Singleton type. Returned by checked_size_product in case the product is not representable in the element type of the input.

See also: StatusInvalidValue

CheckedSizeProduct.checked_size_product — Function

checked_size_product(size_tuple)

In short; a safe product, suitable for computing, e.g., the value of length(dense_array) from the value of size(dense_array). In case everything is as expected, return the value of the product of the elements of the given tuple. Otherwise, return a NamedTuple containing Boolean values with information on the encountered problems.

In more detail; given size_tuple, a nonempty homogeneous tuple of Integer-likes:

Give the name T to eltype(size_tuple) for the purposes of this doc string.

The user must ensure T supports:

Base.Checked.mul_with_overflow

iszero

typemax

<

==

Calculate the product. In case no element is negative, no element is typemax(T) and the product is representable as T, return the product. Otherwise return:

a value of type [StatusInvalidValue]: the input contains a negative element, or an element equal to typemax(T)

a value of type [StatusOverflow]: the product is not representable as T

Throws when given an empty tuple to avoid having to choose a default return type arbitrarily.